moved previous tests and added new one for login
