+++ __tests__/landingPage/Header.test.tsx
@@ -1,6 +1,6 @@
 import { render, act, fireEvent } from "@testing-library/react";
 import "@testing-library/jest-dom";
-import Header from "../components/common/Header";
+import Header from "../../components/common/Header";
 
 describe("Test the Header", () => {
 	it("should change the color mode", async () => {
